pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "jenkins-log-intelligence"
version = "1.2.0"
description = "Jenkins Log Intelligence Engine"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "httpx>=0.27.0",
    "celery[redis]>=5.3.6",
    "redis>=5.0.4",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "pydantic-settings>=2.2.1",
    "slack-sdk>=3.27.1",
    "aiohttp>=3.9.0",
    "anthropic>=0.25.0",
    "scikit-learn>=1.4.2",
    "pyparsing>=3.1.2",
    "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths    = ["app/tests"]
